issue #2 : able to skip duplicates

+++ src/main/java/nl/marisabel/images/ProcessImage.java
@@ -13,6 +13,8 @@ import java.io.IOException;
 import java.nio.file.Files;
 import java.nio.file.Path;
 import java.nio.file.StandardCopyOption;
+import java.text.SimpleDateFormat;
+import java.util.Date;
 import java.util.regex.Pattern;
 import java.util.logging.Logger;
 
@@ -39,20 +41,41 @@ public class ProcessImage {
       String year = dateParts[0];
       String month = dateParts[1];
 
-      // Modify created date to taken date
-      ModifyCreatedDate modifyCreatedDate = new ModifyCreatedDate();
-      modifyCreatedDate.modify(imageFile, dateTime);
-
       // Create the destination folder if it doesn't exist
       Path destinationFolder = Path.of(toFolder, year, month);
       Files.createDirectories(destinationFolder);
 
-      // Move the image file to the destination folder
-      Path destinationFile = destinationFolder.resolve(imageFile.getName());
-      Files.move(imageFile.toPath(), destinationFile, StandardCopyOption.REPLACE_EXISTING);
+      // Rename the image file with the taken date data
+      String originalFileName = imageFile.getName();
+      String extension = originalFileName.substring(originalFileName.lastIndexOf('.'));
+      String newFileName = generateNewFileName(dateTime, extension);
+      Path destinationFile = destinationFolder.resolve(newFileName);
+
+      // Skip renaming if the file already exists in the destination folder
+      if (!Files.exists(destinationFile)) {
+       Files.move(imageFile.toPath(), destinationFile, StandardCopyOption.REPLACE_EXISTING);
+      } else {
+       log.info("Skipped file: " + originalFileName + " - Already exists in the destination folder.");
+      }
      }
     }
    }
+  }
+ }
+
+ private String generateNewFileName(String dateTime, String extension) {
+  SimpleDateFormat dateFormat = new SimpleDateFormat("yyyyMMdd_HHmmss");
+  Date date = parseDate(dateTime);
+  String formattedDate = dateFormat.format(date);
+  return formattedDate + extension;
+ }
+
+ private Date parseDate(String dateTime) {
+  try {
+   return new SimpleDateFormat("yyyy:MM:dd HH:mm:ss").parse(dateTime);
+  } catch (java.text.ParseException e) {
+   log.warning("Error parsing date: " + dateTime);
+   return new Date();
   }
  }
 }
